The beginning of the end for the old dock window layout...
Incorporate the gallery folder history into the gallery widget,
with preferences options to show it or not and where.

Simplify the names shown in the history combo if there is only one
gallery (which is the case at the moment).

imgnamecombo.* -> imagenamecombo.* to match class name
New files kookagallery.{cpp,h}
Eliminate the unused "resource.h"
Add icon for 'photocopy' action
Generate a default description (scan mode and resolution) for new scan parameters

svn path=/branches/work/kooka-kde3/; revision=800988