Update the image saving and format selection dialogue to be a bit more user friendly.
If a file name is being requested, ask for it as part of the
image format dialogue instead of offering two dialogue boxes
in succession.

Make the "Always use the Save Assistant" preferences option
and the format dialogue's "Always use this format..." option
do exactly what they say - they are not equivalent.  The first
is a permanent global option, while the second affects one
type of image only.

svn path=/branches/work/kooka-kde3/; revision=838282