Initial port of Kooka to KDE4.
It compiles, runs and scans, but there is lots more to do to get it working properly.
Some K3/Q3 compatibility classes have been used, and there are some things that are
commented out (with "#ifdef KDE4") at the moment.  Also lots of things to fix marked
with "TODO" in the sources.

- The troublesome dock widgets have been replaced by tabs/splitters.  Layout of these
needs to be configurable, and the main window and splitter sizes need to be saved
properly when exiting and restored on next startup.

- The scan previewer widget layout and auto-select controls are not right.

- Printing and photocopy needs to be ported to use the new Qt-based printing system.

- OCR results spell checking needs to be ported to Sonnet.

- Configuration options and dialogue should be ported to KConfigXT.

- Gallery and thumbnail views need to be updated to the KDE4 way of doing things.

- Thumbnail view setting of preview mode and popup menu does not work.

But it's a start...

svn path=/branches/work/kooka-kde4/; revision=1037174