Move the gallery root setting from libkscan (where it didn't belong and was never used) to the Kooka application (where it belongs).
Standard location for this is now the non-hidden place
<user document path>/KookaGallery.  This new gallery directory will be
created on startup; if the old location also exists then the user will
be reminded where it is and that they can copy the old gallery images
if required.  Not doing this automatically, because of the possibility
of data loss if it goes wrong.

The gallery location or its leaf name can be set with the hidden (no GUI)
configuration option "[Gallery] Location".  This can be an absolute or
a relative path.

svn path=/branches/work/kooka-kde4/; revision=1039356