Rewrite the format dialogue to use MIME types throughout, eliminate duplicates in the available formats (e.g. JPG and JPEG are equivalent).
Big internal reorganisation to eliminate the confusion between a
KimageIO/Qt format name (now the opaque type ImageFormat), a MIME
type and a file extension.

Update the size indicator to give a better indication for small files
and near the thresholds.

Added 'kimageioinfo' utility to show available formats and information.
Put this and 'sanedump' into a new directory 'tools'.

svn path=/branches/work/kooka-kde4/; revision=1044717