Split up kscandevice to separate scanner initialisation and config files (scanglobal), scanner device listing and information (scandevices), and handling the scan device (kscandevice).  Simplifies the API and eliminates a lot of shared data which shouldn't be.
With the access to scanner information that this provides, improve the look
of the scanner selection dialogue (deviceselector, formerly devselector).
The icon displayed for each scanner device is dependent on the SANE device
name or type, as listed in the data file 'scantypes.dat' (installed in the
libkscan application directory).

svn path=/branches/work/kooka-kde4/; revision=1053652