Further porting and cleaning up of the OCR dialogue and engines.
Add an extra tab to the OCR dialogue for the engine information, so as not
to clutter the options page.  Get the version of GOCR and OCRAD correctly.

Another page for an OCR debugging facility (which doesn't do anything yet).

Display the source image preview and information correctly.  Forget about
KFileMetaInfo, just show the basic image information.

Use a progress bar instead of an animated buton to show that OCR is
in progress.

Clean up the signals/slots interface between the engine and dialogue.


svn path=/branches/work/kooka-kde4/; revision=1094278