Prompt for a save filename/format at the beginning of the scan
This is probably a more friendly thing to do than at the end,
if the scan will take a long time - the user can go away and
know that their scan will be saved when it finishes.  Not enough
information may be available from the scanner to ask at this
stage, though, so Kooka will fall back to the old way if this
is the case.

Later, this may be made an option.