Finish the porting of gamma table editing
It should now be possible to edit gamma tables and correctly update
the scanner with them.  At the moment only a single gamma table is
supported, if the scanner has individual gamma tables for each
colour channel then the same gamma table is sent to each.

This should also work (although I am not able to test it) for scanners
with other than 256 gamma table entries, bug 126914.

Also clean up the code and API, and add API documentation to the
gamma table classes.

CCBUG:126914