Fix crash with GOCR: suspicous use of QRegExp::capturedTexts() with a pattern that doesn't have any capture sections defined!
Rewrite the GOCR stdout/stderr handling for better separation
of results, progress and error/debug.

GOCR result text is sent to a temporary file (formerly stdout)
and read when the OCR process is finished.

GOCR standard error is sent to another file (formerly printed via
kDebug(), but nothing else was done with it).

GOCR progress is still sent to stdout - the only thing that should
now be sent to stdout - and used to display progress.  Only really
useful for very long OCR tasks, since GOCR only prints progress
messages at 10 second intervals.

Correct the capture and handling of the GOCR progress output,
and the scanning of the progress line.  Pointless, because nothing
currently uses the progress signal!

Fixed initial value setting problem with KScanSlider which showed
itself in the GOCR setup dialogue page.

BUG:302421