Fix preview area auto-detection apparently not working when the scanner background is white.
It did work really, but the UI was not intuitive because the
threshold slider sets the absolute value of the threshold.  It works
as would be expected for black background - low threshold working
upwards from full black - at the left end of the slider, but for
white background the high threshold working downwards from full
white was at the right end of the slider.  So if the scanner
background is white the user has to not only set the combo
appropriately but also move the tolerance slider to the other end
of its range.  It's not obvious that the latter has to be done.

Now, the tolerance slider works the same way for either black or
white background - low threshold meaning less variation accepted
is at the left.  The slider range is reduced to 50 (from 255) for
easier adjustment within the useful range.  This also simplifies the
area detection code, black and white being treated identically with
only the averaged pixel values being adjusted for white.

Fix initial enabling of auto-detection and its GUI.
Do not turn on auto-detection on initial startup.
Make dust size GUI work (if it is enabled at compile time,
usually it is not).

Rationalise the Previewer class's member variables and names.
Not much point in having a PreviewerPrivate unless everything
is there, it just creates complexity.