Improve the Auto Select user interface and the scan previewer
Move the autoselect controls to an AutoSelectBar below the preview
image, only displayed when auto select is turned on.  Move the
scanner background combo box, along with the previously hidden
dust size slider and a new margin slider to a new AutoSelectDialog
accessed from the bar.

Add an application action (in the scan toolbar by default) to
turn on auto selection and show the bar, along with an icon
for it (currently a bit crude).

Change the threshold and dust size sliders to a KScanSlider to
provide a spin box for the value.  Fix KScanSlider (and other
similar controls) being able to take keyboard focus in this
dialogue and elsewhere.

Centralise the auto selection constants (apart from the config
keys which are only used by Previewer) in a new AutoSelectData
namespace and header file.

Also move the scan size display and file size indicator to the
application status bar, eliminating the waste of horizontal space
caused by it alongside the preview area.

Centralise all status bar handling and item display in a new class
StatusBarManager, instead of being split between Kooka and KookaView.
Gallery image and scan area each have their own status bar indicators,
consistently formatted.