Account for "priority" scanner options
There is nothing special about them as far as SANE is concerned, but
some scanners appear to reset other options to a default when
certain options are changed - e.g. the HP OfficeJet Pro 8620 resets
the resolution to 75 when the scan source is changed.

To get around this, set the "priority" options first and the others,
with their intended values, afterwards.  There is no general way to
detect the options to which this applies, so there is a hardcoded list
which currently only includes "source".

Doing this allows the option creation to be simplified, creating all the
options first and then setting their values, taking account of priority,
afterwards.