Allow for the possibility of list settings changing
These may be either normal list values or resolution, both of which are
displayed in a combo box.  Originally it was assumed that these could be
read on initialisation and would never change, but this does not apply
to, e.g. the HP OfficeJet Pro 8620 where the resolution has a maximum
of 300dpi if the "ADF" or "Duplex" source is selected, but can go up to
1200dpi for the "Flatbed" source.

For these types of control, regenerate the combo box list if necessary
when reloading.