Format dialogue: Filter for allowed as well as recommended types
As well as the already existing recommended image type information for
the known formats, note which image types the formats are allowed for
and filter the format list accordingly.  For some formats it is not
clear what is allowed, so it is assumed that they are allowed for all
image types.

This avoids, for example, suggesting the XBM format for a colour image.