Restore the cut markers and adjacent page indicators
This is Kooka's original scheme, where the numbers printed at the
corners of the pages indicate the index of the adjacent page in that
direction (after fixing a bug in the original affecting the number at
the extreme bottom right of the page).  It isn't totally intuitive
because the adjacent page doesn't have a number on it indicating which
page index it is (and now it doesn't correspond to the order that the
pages come out of the printer either).  Maybe investigate an alternative
later.