Split up the DialogStateSaver into a watcher and a saver
This is so that the saver, which does the actual saving, need not be a
QObject.  The watcher monitors the parent dialogue and calls the saver
at the appropriate time;  the saver can be subclassed and replaced,
but the watcher need not and cannot be.

Therefore the parent dialog can be its own state saver, by inheriting
DialogStateSaver in addition to DialogBase or QDialog and implementing
its own specific saveConfig() and restoreConfig() functions.  By doing
this, instead of using a separate saver class, it has access to its
own internal state.  Otherwise the internals would have to be made
available to the saver class either by access functions or by having it
as a friend, with extra code and complications.