Start of a big rewrite to implement OCR engines as plugins
With the possibilities of other sorts of plugins as well.  In order to
permit that plugins are subclasses of AbstractPlugin, with OCR plugins
being subclasses of AbstractOcrEngine.  Plugins are located by a
KServiceTypeTrader query via their .desktop files, where the name, icon
and description of the engine originate (so the name and description
can be translated there).

In order to be able to link both the main application and a plugin, the
parts of the Kooka application which are also needed by plugins
(application settings, KookaImage, ImageFormat and plugin management)
are compiled into a shared library 'kookacore'.  The OCR base class
(which contains a substantial amount of shared code) is compiled into a
second shared library 'kookaocr'.

Currently only GOCR is implemented and is hardwired for testing.