Make plugin information available, use this in app settings
The plugin information is extracted from the service file and stored in
an AbstractPluginInfo structure in the AbstractPlugin base class.  This
is then available to plugin derived classes without having to pass
around the original KService (or even needing to know that it originally
came from the service desktop file).  The KUIT markup for the comment is
converted at this stage, so there is no need for any other user of the
description to do so.

Implement PluginManager::allPlugins() to enumerate all of the plugins
of a specified type, and currentPlugin() to get the currently loaded
plugin.  All plugin loading and unloading is now handled by
PluginManager.

Use this plugin information for the OCR settings and dialogue, and the
OCR plugin loading and switching.