Centralise OCR process and logging in AbstractOcrEngine
Dealing with the creation, starting and exit status of the process.  The
command line and arguments are assembled by the engine subclass in
createOcrProcess(), and the results are analysed by finishedOcrProcess().
Standard error is automatically logged to a file and displayed if there
is a runtime error.

This allows the protected member variables in AbstractOcrEngine to be
eliminated.