KookaImage: Identify subimages by a URL fragment
So simplifying the logic by eliminating the concept of a 'parent' image.
This allows the subimage detection in ScanGallery::loadImageForItem()
to just generate image URLs, and KookaImage::loadFromUrl() will
automatically handle what previously had to be done by calling
KookaImage::extractNow().

This simplifies KookaImage by eliminating one constructor and not
needing a subimage number to be stored - the URL describes it all.

It also fixes the ScanGallery problem where viewing a subimage, then
unloading it, then viewing it again, would create a set of new subimages
nested under the original subimage (which could be repeated ad infinitum).