ScanImage: Pass around using a shared pointer
Using a QSharedPointer to track image reference counting and lifetime.
So that the image can be allocated once when it is generated (by
scanning, or in a few other cases) and the same copy used by everything
that uses the image. Simplifies a lot of code, and eliminates the
tricky question of exactly who owns the image and when it is safe to
delete it.

The pointer type is ScanImage::Ptr which can be passed around or copied
as required.  It is possible to obtain a plain ScanImage pointer, or a
QImage, from this, but that is only done in some very special cases when
the image lifetime can be assured.