Start to implement "Scan Destination" plugins
So that the scan results can be sent to somewhere other than the
default gallery:  for example, printed (so replacing the "photocopy"
mode that was never ported from KDE3) or to an application.

Move some application source files to the kookacore library so that
they can be used by plugins.  Move the gallery root location
from "kookapref" to a new source file "galleryroot" so that it can
go into the kookacore library.

Send the scan to the destination plugin in KookaView::slotScanStart()
and KookaView::slotNewImageScanned().

Extend the PluginManager to search for destination plugins.

Implement the first destination plugin "gallery" to save the scanned
image in the gallery as before.