GUI for selecting the scan destination
Added to the ScanParams page by the subclassed KookaScanParams,
which also manages selecting and loading the plugin.  The loaded
plugin is retrieved and used by KookaView.

Made the ScanParamsPage a public interface so that KookaScanParams
and the plugin can use it to add GUI widgets.  Added some functions
for controlling the row where widgets are added, so that the plugin
GUI appears in the right place and can consistently use the grid layout.