Always call the plugin's scanStarting() regardless of the "ask" setting
Other plugins may need to get prepared even if the gallery plugin is not
asking for a file name.  Those that need to prompt or interact with the
user can check the setting themselves.

Add a new image type ScanImage::Preview so that KookaView::slotScanStart()
can take account of what is happening.