Make ImgSaver work on a ScanImage::Ptr instead of a QImage
So that it can used the image type derived from the SANE parameters,
set by KScanDevice.  Implement ScanImage::setImageType() to set the
type, and ScanImage::imageType() to do what ImgSaver::findImageType()
would have done if the type is not set.

This means that ImgSaver and also plugins can get the image type
directly from the image, without having to remember it or examine the
image again.