Export plugin: Send the scanned image to an online service
Using the KIPI plugin system to send to any destination that it
supports.

Caution: This works the first time, but selecting another Kooka
destination and then "Online Service" again will crash the
application.  The problem seems to be that KIPI is not cleanly unloaded
even if all of the objects associated with it are deleted, there are
ominous static variables within the KIPI library which are safe in the
sense that they are POD, but do not appear to be reset when their
associated objects are destroyed.  It therefore seems to be not possible
to completely reinitialise or recreate the KIPI system, but retaining
the KIPI objects also seems to have problems in reusing them.

Implemented AbstractDestination::delayedDelete() to eventually delete
temporary files that may need to persist for an indeterminate time.