Export plugin: Use a helper executable to manage KIPI plugins
To avoid the problem of not being able to cleanly tear down and
subsequently recreate the destination plugin and the KIPI plugins
that it uses.

The 'kookakipihelper' program has two modes of operation:  one to
list the available plugins with their action names, icon names and
text, and the second to activate a specific plugin with a list of
image files.  The destination plugin executes the helper in either
a subprocess (with standard output captured) or detached, as
required.

Extend the KIPI interface to be able to handle multiple image files
(although Kooka only uses one at a time).

It works as a solution, but it still brings to mind the observation
as quoted in the "Unix-Haters Handbook" (p.174):  "Treating memory
management through process creation and exit is like medicine treating
illness through living and dying, i.e. it is ignoring the problem."