Do not enter an endless wait when deleting a gallery folder
ScanGallery::slotDeleteItems() would call updateParent() to do exactly
what it says, updating the parent item counts to reflect the removed
item.  However, the parent URL was not being derived correctly which
meant than the item about to be deleted itself was requested to be
updated.  This called FileTreeBranch::slotListerStarted() but due
to the entry then being deleted, FileTreeBranch::slotListerCanceled()
was never called.  The busy cursor was therefore shown continuously.

Correctly derive the parent URL in ScanGallery::updateParent() to update
the correct item and not start a listing for the one about to be deleted.