Add Scan Device dialogue: Show device icons in the type combo
Move the icon lookup for a device or type to ScanDevices, so that
it can be used by this dislaogue and as originally in DeviceSelector.
There need to be two functions: deviceIconName() for when the device
is known and so its type can be obtained from SANE, and typeIconName()
for when only the device type is known.