Add an option to disable the use of the system network proxy
This is necessary for the ESCL ("AirPrint") SANE backend, which
uses libcurl internally to access the scanner over the network.
However, that appears to have problems accessing the device through
a proxy, so it needs to be disabled for scanner operations.  It
can't just be disabled globally for the entire Kooka application,
because it may launch other processes or applications which need
the proxy.

Therefore, if the proxy is disabled in Kooka, the proxy environment
variables are saved and cleared just before a SANE operation that
may need the network, and restored immediately afterwards.

There is a preferences option to continue to use the proxy, if it is
needed and working for use by SANE.  The default for this option is
"off" (disable the proxy), to cover the common case of a local ESCL
device.