KScanCombo: Add scan type icons here instead of in ScanParams
That originally worked for the initial widget creation, as done
in ScanParams::createScannerParams(), but if the list of available
scan modes changes at runtime or when another option changes,
setting the new values clears all the icons.  Seen with the
MFC-J4540 when switching between 'Flatbed' and 'ADF'.

If it is controlling the scan mode, tell the control using
KScanCombo::setUseModeIcons().  If that has been set, set the icons
corresponding to the string values in KScanCombo::setList().