Ask user before scan depending on ADF mode
In ADF mode it is necessary to delay asking the user for a file name
and/or type until after sane_start() has been done, in order that an
empty ADF can be detected.  However, this introduces a delay, while the
scanner starts up, before the dialogue.  So only delay in ADF mode,
and ask as early as possible otherwise.