Share destination: Handle a batch scan in one operation
Save the individual images to temporary files, then after the
end of the batch send them all to Purpose in one share job.

TODO: Assuming that this is possible and sensible to do for all
share destinations.  May need a check box to select whether to
share as a batch or individually.