Remove the "Ask for filename before/after scanning" option
Originally the file name and format was requested after scanning.
Commit 21e0bd99 moved the request to happen at the beginning of the
scan instead, because this seemed to be the better option where the
scan would take a long time.  Commit 03d01e55 then introduced a
preferences option for this, with the default still being to ask before.

Asking before still seems to be the best option from a workflow point
of view, and will simplify the code and interaction for multiple scans.
So the option to ask afterwards is removed.