Refactor the scanner parameters and options storage and access
Rename the groups in the scanner configuration file so that startup
parameters are saved on a per-scanner basis and there is a clear
distinction between scanner parameters and user saved presets.

Move the logic for finding a configuration group from ScanParams to
KScanOptSet (because it is mostly used there), and add a 'category'
parameter to specify the category of settings which are required -
scanner parameters, general scanner options, or user presets.

Add a version number to the configuration file and perform a migration
at ScanGlobal initialisation time if necessary.

Do not require a KScanOptSet description to be provided for saving and
do not return it via listSavedSets() (formerly "readList"), there are
accessor functions for this.