Show an information message if multiple scanning is enabled
Now that the multiple scan options are persistent.  It is only shown
the first time that the application is started, or a new scanner is
selected, to avoid inadvertently starting off a multiple scan operation.
The message can be closed, or it will disappear automatically when:

- the user opens the "Multiple Scan" dialogue
- they initiate a preview or a scan

since it is assumed that at that stage they know what they are
intending to do.