Application destination: Declare capabilities and handle no-batch more intuitively
As originally implemented, if batching was disabled for a multiple scan
then all of the images would be accumulated until the end of the batch
as usual, then a separate application would be launched for each of them.
Now, if batching is disabled the application is launched for each image
as soon as it has been scanned.