Let a destination plugin cancel or not accept an image
By allowing AbstractDestination::imageScanned() to return a failure
status.  This is intended to be used where there is a user interaction,
which must happen there because the image needs to be available,
which the user cancels (test case: the 'print' destination).  This is
treated in the same way as a cancel during scanning.