KookaPrint: Accept multiple images for a single print job
Which needed quite a bit of reworking, splitting up printImage()
into a preamble startPrint(), printImage() with an image supplied
and endPrint() to finish.  A reference image is set to provide the
size and resolution, but which is not actually printed.  Any number
of images can then be given to combine into a single job, which
are assumed to be compatible with the reference.