Save destination: Do not add "All files" to the filter list
This was added in commit 83641e69, with the intention being to allow
the user to be able to save to an image format that does not appear
in the filter list.  However, including this filter entry for saving
interferes with the "Automatically add file extension" option -
selecting a filter and then starting to type a file name without
any extension (yet) changes the filter to "All files" and suggests the
partial file name as the extension.

If the user really wants to enter a file name that does not correspond
to any of the filters, they can turn off the check box and then type
the full file name.