Multipage destination: Implement our own page setup dialogue
Only a few of the QPageSetupDialog options are actually needed,
and it is not possible to preset it with the current settings.
Just remember the configured paper size instead (which also implies
its orientation), and use a new MultipageOptionsDialog with only
the paper size (obtained from libpaper) and orientation options.