Ensure that the scan area indicator is initialised on startup
More complicated than would appear, because originally on startup the
initial KScanDevice was set up during the KookaView constructor which
in turn was called early on in the Kooka constructor.  This happened
before the settings had been loaded or the signal connections from the
KookaView to the status bar indicators had been connected.

The KScanDevice is now set up in the new function KookaView::initScanDevice()
which is called when the KookaView and Kooka objects are fully constructed.
This ensures that the status bar has been set up and its signals connected
before initialising the scan area.

Return the actual rectangle that was set by ScanParams::applyRect() so
that ScanParams::initStartupArea() can emit it via the newCustomScanSize()
signal.  This ensures that it reaches the Previewer and status bar via
the signal connections made above.