KScanDevice: Generate an image corresponding to the scanned image type
That is, generate a "Mono" or "Grayscale" image immediately if the
scan parameters indicate so, instead of converting the image to the
requested image format afterwards.  Originally "Grayscale" was
generated as a LowColor image with a greyscale palette, but QImage
now supports a greyscale format directly.

SANE will still never produce a LowColour image, so that is generated
as RGB and then down converted if necessary.