Implement the ""Delay/wait before first scan" option
Not so simple as just moving the delay to the start of the loop,
because its prompting will be slightly different and it also needs
to happen after any prompting done by the destination plugin.

Implement KScanDevice::delayWait() to do the prompting and check
the result.

Extend ContinueScanDialog to take an 'isFirst' flag and give
different messages and buttons accordingly.