ContinueScanDialog: Accept and action the "can't cancel" option
Originally derived from the destination plugin's capabilities, this
indicates that any previous scans will already have been processed
and hence "Cancel" will have the same effect as "Finish".  To make
it clear to the user, only show "Finish" in this case.

Set that capability for those destination plugins where it is
appropriate.